Split large TUSClientTests into separate suites.
